% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_breachLC_fillWang_D8.R
\name{stream_breachLC_fillWang_D8}
\alias{stream_breachLC_fillWang_D8}
\title{Title}
\usage{
stream_breachLC_fillWang_D8(input_dem_path, output_folder_path, fill_dist, x)
}
\arguments{
\item{x}{}
}
\description{
Title
}
\examples{
stream_breachLC_fillWang_D8("D:/04_DYDAKTYKA/01_SEKCJA_GEO/_Projekt_Hydro_BUL/Hydro_20240229/NMT_clip.tif", "D:/04_DYDAKTYKA/01_SEKCJA_GEO/_Projekt_Hydro_BUL/", 5, 2500)
}
