% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydR-package.R
\docType{package}
\name{hydR-package}
\alias{hydR}
\alias{hydR-package}
\title{hydR: Package for creating hydrological network inside forest}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\keyword{internal}
